destinations/webhooks/python/requirements.txt
Flask==2.1.0
python-dotenv==0.15.0